+++ PlaceFeedsServices/MeetupService/IMeetupService.cs
@@ -4,6 +4,6 @@ namespace PlaceFeedsServices.MeetupService
 {
     public interface IMeetupService
     {
-        Task<string> GetMeetupData(float latitude, float longitude);
+        Task<string> GetMeetupData(double latitude, double longitude);
     }
 }
